chore(editor): Remove ACE
