+++ package.json
@@ -217,7 +217,6 @@
     "PrettyCSS": "0.3.19",
     "array-to-sentence": "2.0.0",
     "bowser": "2.9.0",
-    "brace": "0.11.1",
     "classnames": "2.2.6",
     "codemirror": "5.52.2",
     "core-js": "3.6.5",
