+++ src/components/EditorsColumn.jsx
@@ -17,7 +17,6 @@ export default function EditorsColumn({
   errors,
   resizableFlexGrow,
   resizableFlexRefs,
-  implementation,
   isFlexResizingSupported,
   isTextSizeLarge,
   requestedFocusedLine,
@@ -62,7 +61,6 @@ export default function EditorsColumn({
             requestedFocusedLine={requestedFocusedLine}
             source={currentProject.sources[language]}
             textSizeIsLarge={isTextSizeLarge}
-            useCodeMirror={implementation === 'codemirror'}
             onAutoFormat={onAutoFormat}
             onInput={handleInputForLanguage(language)}
             onReady={partial(onEditorReady, language)}
@@ -98,7 +96,6 @@ export default function EditorsColumn({
 EditorsColumn.propTypes = {
   currentProject: PropTypes.object.isRequired,
   errors: PropTypes.object.isRequired,
-  implementation: PropTypes.oneOf(['ace', 'codemirror']),
   isFlexResizingSupported: PropTypes.bool.isRequired,
   isTextSizeLarge: PropTypes.bool.isRequired,
   requestedFocusedLine: PropTypes.instanceOf(EditorLocation),
@@ -115,6 +112,5 @@ EditorsColumn.propTypes = {
 };
 
 EditorsColumn.defaultProps = {
-  implementation: 'ace',
   requestedFocusedLine: null,
 };
