+++ src/containers/Console.js
@@ -17,7 +17,6 @@ import {
   getCurrentConsoleInputValue,
   getCurrentProjectKey,
   getHiddenUIComponents,
-  getRemoteConfig,
   getRequestedFocusedLine,
   isCurrentProjectSyntacticallyValid,
   isTextSizeLarge,
@@ -29,7 +28,6 @@ function mapStateToProps(state) {
     currentProjectKey: getCurrentProjectKey(state),
     history: getConsoleHistory(state),
     currentInputValue: getCurrentConsoleInputValue(state),
-    useCodeMirror: getRemoteConfig(state).get('editor', 'ace') === 'codemirror',
     isOpen: !getHiddenUIComponents(state).includes('console'),
     isTextSizeLarge: isTextSizeLarge(state),
     requestedFocusedLine: getRequestedFocusedLine(state),
